K Most occurring elements(K_most_frequent_elements)
Given an array arr[] of N integers in which elements may be repeating several times. Also, a positive number K is given and the task is to find sum of total frequencies of K most occurring elements

For each testcase, print the sum of total frequencies of K most occurring elements in the given array.